1.37.0 stuff

+++ src/UI/PanelScreenViewController.cpp
@@ -14,7 +14,7 @@ namespace SliceDetails
         auto layouts = get_transform()->GetComponentsInChildren<UnityEngine::UI::HorizontalLayoutGroup*>();
         for (int i = 0; i < 12; i++)
         {
-            auto* layout = layouts[(int)(i/4)];
+            auto* layout = layouts[i/4];
             panelImages[i] = PanelUI::New_ctor(i);
             BSML::parse_and_construct(IncludedAssets::GridInfoUI_bsml, layout->get_transform(), panelImages[i]);
         }
